Chart Redesign - Phase 1 - Move to Tab Controller
